fix(console): Add ClearLine
